refactor admin test packages structure
